:iiam:
This is the test build I made back in June for SQLite.  It also contains a bunch of other changes that I did document at one time but I currently do not have on me.  This branch should be considered a pre-pre-pre-pre-alpha test build and in all honesty probably does not work.  However I have been promising to upload code for quite some time now and I have now done so.  I will begin working with this code to get it into a working condition if it appears to not be and then it will be moved from a test build to a stable build and finally to trunk.  No time frame is given for this.