Hovermenu is being cached.  I am trying to get to the bottom of it all.
This is a huge mess and will likely get reverted, because this code is GARBAGE but it's my work in progress.  See what you can do with this.  Hopefully something.  

I guess this has always been happening but we just never noticed?
Don't rebuild menu with this (so don't post or anything if you have a test build) because I've got some hackish stuff going on to try to figure this out.